project/aidlc-docs/construction/unit3_card_creation/src/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23
alembic==1.12.1

# Redis & Messaging
redis[hiredis]==5.0.1
boto3>=1.34.0

# External APIs
aiohttp==3.9.1
google-api-python-client==2.108.0

# Image Processing
Pillow==10.1.0

# Validation & Serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
